App/requirements.txt
Flask>=2.2.2
Werkzeug>=2.2.0
kafka-python==2.0.2
mysql-connector-python==8.0.33
pandas==1.5.3
numpy==1.21.6
joblib
scikit-learn==1.2.2
sqlalchemy


